Repo, Entity 수정
