--- app/src/main/java/com/example/myapplication/WarehouseRepository.kt
@@ -1,24 +1,0 @@
-package com.example.myapplication
-
-import android.graphics.Bitmap
-import androidx.lifecycle.MutableLiveData
-import com.google.firebase.database.DataSnapshot
-import com.google.firebase.database.DatabaseError
-import com.google.firebase.database.ValueEventListener
-import com.google.firebase.database.ktx.database
-import com.google.firebase.ktx.Firebase
-
-interface WarehouseRepository {
-
-    suspend fun deleteItem(name: String)
-
-    suspend fun getItem(name: String): List<WarehouseEntity>
-
-    suspend fun addItem(
-        name: String,
-        rentalUser : String,
-        totalItem: Int,
-        currentItem: Int,
-        rentalState: Boolean
-    )
-}
